Add query ID to raised error
We're getting an EOFError thousands of times a day and ClickHouse has
asked us to give them query IDs to that they can look into it. I
previously put in functionality to add the query ID to the raised error,
but the one place I missed is the one place throwing all the errors.

This sets a query ID that will be used in the HTTP(S) request and will
be in the error message where this is being raised.